move block->epoch APIs to EpochManagerHandle (#7665)
Part of https://github.com/near/nearcore/issues/6910, depends on https://github.com/near/nearcore/pull/7660

Must be reviewed per commit.
